WFM
If I run
npm i; ts-node index.ts

This script will run the openvpn client on mu mac to see if it is
connected.

If not, it will then retrieve my credentials from the system password
vault (keychain) and then pass them to the openvpn client for
reconnection.

There is no way to set the keychain currently